+++ backend/service/supplierService.js
@@ -4,14 +4,15 @@ const {sequelize} = require('../models');
 
 
 class SupplierService{
-    static async createSupplier(data){
-        const transaction = await sequelize.transaction();
+    static async createSupplier(data, transaction=undefined){
+        let flag = false
+        if(!transaction){flat = true;transaction = await sequelize.transaction();}
         try{
             const supplier = await SupplierDAO.create(data, transaction);
-            await transaction.commit();
+            if(flag){await transaction.commit();}
             return supplier;
         }catch(error){
-            await transaction.rollback();
+            if(flag){await transaction.rollback();}            
             console.error('Service Error (createSupplier):', error.message);
             throw error;
         }
@@ -31,20 +32,22 @@ class SupplierService{
      * @param {number} id - Supplier id
      * @returns {supplier|exception}  Throws exception if can not find any match with the id
      */
-    static async findById(id, include, transaction=undefined){
+    static async findById(id, include=undefined, transaction=undefined){
+        let flag = false;
+        if(!transaction){flag=true; transaction = await sequelize.transaction();}
         try{
             const associationMap = {
                 d: 'Donation',
                 dd: 'DonationDetails'
             }
-            include = include.map(item=> associationMap[item]|| undefined)
+            console.log('findbyid supplier service', id)
+            include = include?.map(item=> associationMap[item]|| undefined)  
             const supplier = await SupplierDAO.findById(id, include, transaction);
-            if(supplier){
-                return supplier;
-            }else{
-                 throw new Error('Supplier Not Found.')
-            }
+            if(flag) await transaction.commit();
+            return supplier
+           
         }catch(error){
+            if(flag) await transaction.rollback()
             console.log('SupplierService Error (findById):', error.message)
             throw error
         }
@@ -57,6 +60,7 @@ class SupplierService{
         try{
             // add validation for do not change the id!
             const oldSupplier = await this.findById(id);
+            if(!oldSupplier){throw new Error('Supplier Not Found.');}
             const newSupplier = await SupplierDAO.update(oldSupplier, data, transaction);
             await transaction.commit();
             return newSupplier;
@@ -80,6 +84,7 @@ class SupplierService{
         const transaction = await sequelize.transaction();
         try{
             const supplier = await this.findById(id);
+            if(!supplier){throw new Error('Supplier Not Found!')}
             const supplierDeleted = SupplierDAO.deleteById(id, transaction);
             await transaction.commit();
             return supplierDeleted;
